requirements.txt
pandas==1.3.2
python_telegram_bot==20.0b0
pytz==2021.1
telegram==0.0.1
python-dotenv==0.21.0
openpyxl==3.0.10